requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
accelerate>=0.20.0

# Text Processing
nltk>=3.8.1
spacy>=3.6.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Interface
streamlit>=1.25.0
gradio>=3.35.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0

# Additional NLP
sentence-transformers>=2.2.2
textstat>=0.7.3
wordcloud>=1.9.2

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0